externalize Entity, remove macro function from nexus
